auto gallery project
Auto Gallery Project is a web application that allows you to add and delete vehicles and stores data in Local Storage. In this project, basic functionalities such as adding tools to the interface, creating error messages, adding and deleting tools to local storage have been implemented. Vehicle management is provided with a user-friendly interface.